refactor: migrate test conversation-store imports to source modules
