+++ assistant/src/__tests__/conversation-routes-guardian-reply.test.ts
@@ -76,7 +76,7 @@ mock.module("../runtime/confirmation-request-guardian-bridge.js", () => ({
   bridgeConfirmationRequestToGuardian: async () => undefined,
 }));
 
-mock.module("../memory/conversation-store.js", () => ({
+mock.module("../memory/conversation-crud.js", () => ({
   addMessage: (
     conversationId: string,
     role: string,
