+++ assistant/src/__tests__/memory-regressions.test.ts
@@ -99,7 +99,7 @@ import {
   getConversationMemoryScopeId,
   messageMetadataSchema,
   provenanceFromTrustContext,
-} from "../memory/conversation-store.js";
+} from "../memory/conversation-crud.js";
 import { getDb, initializeDb, resetDb } from "../memory/db.js";
 import { selectEmbeddingBackend } from "../memory/embedding-backend.js";
 import {
