+++ assistant/src/__tests__/relay-server.test.ts
@@ -202,7 +202,7 @@ import {
   createInboundSession,
   createVerificationSession,
 } from "../memory/channel-guardian-store.js";
-import { addMessage, getMessages } from "../memory/conversation-store.js";
+import { addMessage, getMessages } from "../memory/conversation-crud.js";
 import { getDb, initializeDb, resetDb, resetTestTables } from "../memory/db.js";
 import { createInvite } from "../memory/invite-store.js";
 import { conversations } from "../memory/schema.js";
