+++ assistant/src/__tests__/runtime-attachment-metadata.test.ts
@@ -59,8 +59,8 @@ import {
   uploadAttachment,
 } from "../memory/attachments-store.js";
 import * as channelDeliveryStore from "../memory/channel-delivery-store.js";
+import * as conversationStore from "../memory/conversation-crud.js";
 import { getOrCreateConversation } from "../memory/conversation-key-store.js";
-import * as conversationStore from "../memory/conversation-store.js";
 import { getDb, initializeDb, resetDb, resetTestTables } from "../memory/db.js";
 import { RuntimeHttpServer } from "../runtime/http-server.js";
 import * as pendingInteractions from "../runtime/pending-interactions.js";
