+++ assistant/src/__tests__/session-abort-tool-results.test.ts
@@ -91,7 +91,7 @@ mock.module("../memory/admin.js", () => ({
 // Track all messages persisted to DB
 let persistedMessages: Array<{ role: string; content: string }> = [];
 
-mock.module("../memory/conversation-store.js", () => ({
+mock.module("../memory/conversation-crud.js", () => ({
   getConversationThreadType: () => "default",
   setConversationOriginChannelIfUnset: () => {},
   updateConversationContextWindow: () => {},
@@ -112,13 +112,16 @@ mock.module("../memory/conversation-store.js", () => ({
     totalEstimatedCost: 0,
   }),
   createConversation: () => ({ id: "conv-1" }),
-  listConversations: () => [],
   addMessage: (_convId: string, role: string, content: string) => {
     persistedMessages.push({ role, content });
     return { id: `msg-${persistedMessages.length}` };
   },
   updateConversationUsage: () => {},
   updateConversationTitle: () => {},
+}));
+
+mock.module("../memory/conversation-queries.js", () => ({
+  listConversations: () => [],
 }));
 
 mock.module("../memory/retriever.js", () => ({
