+++ assistant/src/__tests__/session-agent-loop.test.ts
@@ -119,7 +119,7 @@ mock.module("../hooks/manager.js", () => ({
   }),
 }));
 
-mock.module("../memory/conversation-store.js", () => ({
+mock.module("../memory/conversation-crud.js", () => ({
   getConversationThreadType: () => "default",
   setConversationOriginChannelIfUnset: () => {},
   updateConversationUsage: () => {},
