+++ assistant/src/__tests__/session-init.benchmark.test.ts
@@ -173,13 +173,10 @@ mock.module("../config/loader.js", () => ({
 
 // Additional mocks required for Session constructor and end-to-end tests
 
-mock.module("../memory/conversation-store.js", () => ({
+mock.module("../memory/conversation-crud.js", () => ({
   addMessage: () => ({ id: "msg-1" }),
-  countConversations: () => 0,
   getMessages: () => [],
-  listConversations: () => [],
   getConversation: () => null,
-  getLatestConversation: () => null,
   createConversation: () => ({
     id: "bench-conv",
     title: "Bench",
@@ -193,15 +190,21 @@ mock.module("../memory/conversation-store.js", () => ({
   getConversationOriginInterface: () => null,
   getConversationThreadType: () => "standard",
   getConversationMemoryScopeId: () => "default",
-  isLastUserMessageToolResult: () => false,
   provenanceFromTrustContext: () => ({}),
   relinkAttachments: () => 0,
-  searchConversations: () => [],
   setConversationOriginChannelIfUnset: () => {},
   updateConversationContextWindow: () => {},
   updateConversationTitle: () => {},
   updateConversationUsage: () => {},
   updateMessageContent: () => {},
+}));
+
+mock.module("../memory/conversation-queries.js", () => ({
+  countConversations: () => 0,
+  listConversations: () => [],
+  getLatestConversation: () => null,
+  isLastUserMessageToolResult: () => false,
+  searchConversations: () => [],
 }));
 
 mock.module("../hooks/manager.js", () => ({
