+++ assistant/src/__tests__/session-workspace-cache-state.test.ts
@@ -78,7 +78,7 @@ mock.module("../security/secret-allowlist.js", () => ({
   resetAllowlist: () => {},
 }));
 
-mock.module("../memory/conversation-store.js", () => ({
+mock.module("../memory/conversation-crud.js", () => ({
   getConversationThreadType: () => "default",
   setConversationOriginChannelIfUnset: () => {},
   provenanceFromTrustContext: () => ({
@@ -103,6 +103,9 @@ mock.module("../memory/conversation-store.js", () => ({
   updateConversationContextWindow: () => {},
   deleteMessageById: () => ({ segmentIds: [], orphanedItemIds: [] }),
   deleteLastExchange: () => 0,
+}));
+
+mock.module("../memory/conversation-queries.js", () => ({
   isLastUserMessageToolResult: () => false,
 }));
